Upgrade Gradle to 6.8.3, Kotlin to 1.4.31 and Ktor to 1.5.2
