+++ gradle.properties
@@ -1,1 +1,2 @@
-kotlin.version=1.3.70
+kotlin.version=1.4.31
+ktor.version=1.5.2
